// ulp-portal/src/main/portal-fe/src/pages/Session/Session.tsx
/*
 * ulp-portal - United Login Platform
 * Copyright (c) 2022-Present Frank Zhang
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import { QuestionCircleOutlined } from '@ant-design/icons';
import type { ActionType, ProColumns } from '@ant-design/pro-components';
import { PageContainer, ProTable } from '@ant-design/pro-components';
import { App, Badge, Popconfirm, Space, Table } from 'antd';
import { useRef } from 'react';
import type { SessionList } from './data.d';
import { getSessions, removeSessions } from './service';
import { useIntl } from '@@/plugin-locale';

export default () => {
  const intl = useIntl();
  const { message } = App.useApp();
  const actionRef = useRef<ActionType>();
  /**
   * columns
   */
  const columns: ProColumns<SessionList>[] = [
    {
      title: intl.formatMessage({ id: 'pages.other.session.column.session_id' }),
      dataIndex: 'sessionId',
      ellipsis: true,
      copyable: true,
      align: 'center',
      width: 150,
      fixed: 'left',
      search: false,
    },
    {
      title: intl.formatMessage({ id: 'pages.other.session.column.username' }),
      dataIndex: 'username',
      width: 120,
      ellipsis: true,
      fieldProps: { autoComplete: 'off' },
    },
    {
      title: intl.formatMessage({ id: 'pages.other.session.column.ip' }),
      width: 130,
      dataIndex: 'ip',
      ellipsis: true,
      search: false,
      render: (_, { geoLocation }) => {
        return geoLocation?.ip ? (
          <Badge status={'success'} text={geoLocation?.ip} />
        ) : (
          <Badge status={'error'} text={intl.formatMessage({ id: 'app.unknown' })} />
        );
      },
    },
    {
      title: intl.formatMessage({ id: 'pages.other.session.column.geo_location' }),
      width: 120,
      ellipsis: true,
      search: false,
      render: (_, { geoLocation }) => {
        return (
          (geoLocation &&
            geoLocation?.countryName &&
            geoLocation?.provinceName &&
            geoLocation?.cityName &&
            geoLocation?.countryName +
              '.' +
              geoLocation?.provinceName +
              '.' +
              geoLocation?.cityName) ||
          '-'
        );
      },
    },
    {
      title: intl.formatMessage({ id: 'pages.other.session.column.device_type' }),
      ellipsis: true,
      width: 110,
      search: false,
      render: (_, { userAgent }) => {
        return userAgent?.deviceType || '-';
      },
    },
    {
      title: intl.formatMessage({ id: 'pages.other.session.column.platform' }),
      ellipsis: true,
      width: 110,
      search: false,
      render: (_, { userAgent }) => {
        if (userAgent?.platform && userAgent?.platformVersion) {
          return userAgent?.platform + ' ' + userAgent?.platformVersion || '-';
        }
        return '-';
      },
    },
    {
      title: intl.formatMessage({ id: 'pages.other.session.column.browser' }),
      ellipsis: true,
      width: 110,
      search: false,
      render: (_, { userAgent }) => {
        if (userAgent?.browser && userAgent?.browserMajorVersion) {
          return userAgent?.browser + ' ' + userAgent?.browserMajorVersion || '-';
        }
        return '-';
      },
    },
    {
      title: intl.formatMessage({ id: 'pages.other.session.column.login_time' }),
      ellipsis: true,
      dataIndex: 'loginTime',
      align: 'center',
      width: 200,
      valueType: 'dateTime',
      search: false,
    },
    {
      title: intl.formatMessage({ id: 'pages.other.session.column.last_request' }),
      ellipsis: true,
      dataIndex: 'lastRequest',
      width: 170,
      valueType: 'dateTime',
      search: false,
    },
    {
      title: intl.formatMessage({ id: 'app.option' }),
      valueType: 'option',
      width: 60,
      align: 'center',
      fixed: 'right',
      render: (text: any, { sessionId }) => {
        return [
          <Popconfirm
            title={intl.formatMessage({ id: 'pages.other.session.delete.confirm' })}
            placement="bottomRight"
            icon={
              <QuestionCircleOutlined
                style={{
                  color: 'red',
                }}
              />
            }
            onConfirm={async () => {
              const { success } = await removeSessions(sessionId);
              if (success) {
                message.success(intl.formatMessage({ id: 'app.operation_success' }));
                actionRef.current?.reload();
              }
            }}
            okText={intl.formatMessage({ id: 'app.yes' })}
            cancelText={intl.formatMessage({ id: 'app.no' })}
            key="offline"
          >
            <a target="_blank" key="remove" style={{ color: 'red' }}>
              {intl.formatMessage({ id: 'pages.other.session.delete' })}
            </a>
          </Popconfirm>,
        ];
      },
    },
  ];

  return (
    <PageContainer content={intl.formatMessage({ id: 'pages.other.session.desc' })}>
      <ProTable<SessionList>
        search={false}
        options={{
          density: false,
          setting: true,
          reload: true,
          search: false,
        }}
        cardProps={{ style: { overflow: 'auto' } }}
        scroll={{ x: 700 }}
        columns={columns}
        rowSelection={{
          selections: [Table.SELECTION_ALL, Table.SELECTION_INVERT],
        }}
        tableAlertRender={({ selectedRowKeys, onCleanSelected }) => (
          <Space size={24}>
            <span>
              {intl.formatMessage({ id: 'app.selected' })} {selectedRowKeys.length}{' '}
              {intl.formatMessage({ id: 'app.item' })}
              <a style={{ marginLeft: 8 }} onClick={onCleanSelected}>
                {intl.formatMessage({ id: 'app.deselect' })}
              </a>
            </span>
          </Space>
        )}
        tableAlertOptionRender={(rowSelection) => {
          return (
            <Space size={16}>
              <Popconfirm
                title={intl.formatMessage({ id: 'pages.other.session.batch-delete.confirm' })}
                placement="bottomRight"
                icon={
                  <QuestionCircleOutlined
                    style={{
                      color: 'red',
                    }}
                  />
                }
                onConfirm={async () => {
                  const { success } = await removeSessions(rowSelection.selectedRowKeys.join(','));
                  if (success) {
                    message.success(intl.formatMessage({ id: 'app.operation_success' }));
                    rowSelection.onCleanSelected();
                    actionRef.current?.reload();
                  }
                }}
                okText={intl.formatMessage({ id: 'app.yes' })}
                cancelText={intl.formatMessage({ id: 'app.no' })}
                key="offline"
              >
                <a target="_blank" key="remove" style={{ color: 'red' }}>
                  {intl.formatMessage({ id: 'pages.other.session.batch-delete' })}
                </a>
              </Popconfirm>
            </Space>
          );
        }}
        actionRef={actionRef}
        pagination={{ defaultPageSize: 10 }}
        request={getSessions}
        rowKey="sessionId"
        dateFormatter="string"
      />
    </PageContainer>
  );
};
